Fix inconsistent namings
Functions returning senders should be named to indicate that
the returned object needs to be started for the action to initiate.

For instance `client.connect_socket` leads one to believe that
the socket connection procedure will begin immediately. In reality
it returned a sender that had to be started for the socket to connect.

This brings the client-facing interface in line since we already had
publisher and reusable_publisher, which already adhered to this
convention.

`client::connect_socket` -> `client::socket_connector`
`client::handshake` -> `client::handshaker`